requirements.txt
Bio==1.8.1
biopython==1.85
fastapi==0.128.0
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
SQLAlchemy==2.0.45
alembic
pydantic-settings
pytest
httpx
python-json-logger
PyYAML

